Refactor : rename and re-argument IProfileSelectionRequestHandler method.
Changes the handler interface to use a handler method,
to take HttpServletRequest rather than Session as argument.

Updates the unit test for SessionAttributeProfileMapperImpl to reflect the API change,
simplifies the test to have less implementation knowledge about where storage is happening in session,
and adjusts the test to use TestDox style method names.